Update and rename search.py to search_handler.py. Update functions in order to read data DIRECT from DB.
